stock_alert_portfolio/requirements.txt
# Stock Alert & Portfolio Management System

# 核心依赖
pandas>=1.5.0
numpy>=1.23.0
requests>=2.28.0
python-dotenv>=1.0.0
click>=8.0.0

# 数据库 (使用不需要编译的版本)
sqlalchemy>=1.4.0

# 定时任务
apscheduler>=3.10.0

# 可视化
plotly>=5.14.0
matplotlib>=3.6.0

# Web 框架
streamlit>=1.28.0

# 数据源 (可选)
yfinance>=0.2.0
tushare>=1.2.0
